% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{active_profiles}
\alias{active_profiles}
\title{Get active Quarto profiles}
\usage{
active_profiles(default = character())
}
\arguments{
\item{default}{Value returned when `QUARTO_PROFILE` is not set.}
}
\value{
A character vector of active profiles.
}
\description{
Reads active profiles from the `QUARTO_PROFILE` environment variable.
Multiple profiles separated by commas are supported.
}
