% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{quarto_document_dir}
\alias{quarto_document_dir}
\title{Get the current Quarto document directory}
\usage{
quarto_document_dir(default = ".")
}
\arguments{
\item{default}{Value returned when `QUARTO_DOCUMENT_PATH` is not set.}
}
\value{
A directory path.
}
\description{
Uses `QUARTO_DOCUMENT_PATH` when available.
}
