% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartoHelpers-package.R
\docType{package}
\name{QuartoHelpers-package}
\alias{QuartoHelpers}
\alias{QuartoHelpers-package}
\title{QuartoHelpers: Helper functions for Quarto Cli}
\description{
Functions for Quarto build conditional content based on profiles. Profiles can be used with _quarto-profile.yml.
}
\author{
\strong{Maintainer}: José Andrés Flores García \email{andres.flores@uaslp.mx}

}
\keyword{internal}
