% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{resolve_quarto_path}
\alias{resolve_quarto_path}
\title{Resolve a path relative to a Quarto context}
\usage{
resolve_quarto_path(
  path,
  root = c("project", "document", "none"),
  project_root = NULL,
  document_dir = NULL,
  must_exist = FALSE
)
}
\arguments{
\item{path}{Relative or absolute path.}

\item{root}{One of `"project"`, `"document"` or `"none"`.}

\item{project_root}{Quarto project root. Used when `root = "project"`.}

\item{document_dir}{Document directory. Used when `root = "document"`.}

\item{must_exist}{Logical. If `TRUE`, error when the resolved path does not
exist.}
}
\value{
A resolved path.
}
\description{
Resolve a path relative to a Quarto context
}
